Move LabelMeta to libs/map/types.ts, remove ref attribute
